pundit authorization

+++ app/controllers/words_controller.rb
@@ -23,9 +23,12 @@ class WordsController < ApplicationController
 
   def show; end
 
-  def edit; end
+  def edit
+    authorize @word
+  end
 
   def update
+    authorize @word
     if @word.update(word_params)
       redirect_to(word_path(@word))
     else
@@ -34,6 +37,7 @@ class WordsController < ApplicationController
   end
 
   def destroy
+    authorize @word
     if @word.destroy
       redirect_to(words_path)
     else
